backend/requirements.txt
flask
flask-cors
python-chess
torch
numpy
requests
gunicorn
